move class and methods for building batch submission instructions for a flowcell run into new module 'flowcellsubmit'
